fixed int overrun in memory tracker
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@2274 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinxstd.h
@@ -176,7 +176,7 @@ void *			operator new [] ( size_t iSize, const char * sFile, int iLine );
 int				sphAllocsCount ();
 
 /// total allocated bytes
-int				sphAllocBytes ();
+int64_t			sphAllocBytes ();
 
 /// get last alloc id
 int				sphAllocsLastID ();
